Updated batchAsset, assign tag menu
Added Batch Import for rig animation data, allowing for text scroll list to recognised .ma and .mb files as well. def batchAsset to be used in conjunction with the export anim data tool if want to import anim data 

Need to add ctrl as assign tag to all controllers to be able to export animation data out.